requirements.txt
# Core dependencies for data pipeline

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # For parquet file support

# Data ingestion
requests>=2.31.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

